CR-998 create generic retryable cloud data store
